+++ hutool-poi/src/main/java/cn/hutool/poi/excel/ExcelUtil.java
@@ -7,12 +7,9 @@ import java.io.InputStream;
 import java.io.OutputStream;
 import java.io.PushbackInputStream;
 
-import org.apache.poi.hssf.usermodel.HSSFWorkbook;
 import org.apache.poi.poifs.filesystem.FileMagic;
 import org.apache.poi.ss.usermodel.Sheet;
 import org.apache.poi.ss.usermodel.Workbook;
-import org.apache.poi.ss.usermodel.WorkbookFactory;
-import org.apache.poi.xssf.usermodel.XSSFWorkbook;
 
 import cn.hutool.core.io.FileUtil;
 import cn.hutool.core.io.IORuntimeException;
@@ -21,7 +18,6 @@ import cn.hutool.core.util.StrUtil;
 import cn.hutool.poi.excel.sax.Excel03SaxReader;
 import cn.hutool.poi.excel.sax.Excel07SaxReader;
 import cn.hutool.poi.excel.sax.handler.RowHandler;
-import cn.hutool.poi.exceptions.POIException;
 
 /**
  * Excel工具类
@@ -207,7 +203,7 @@ public class ExcelUtil {
 	public static ExcelReader getReader(File bookFile, String sheetName) {
 		return new ExcelReader(bookFile, sheetName);
 	}
-	
+
 	/**
 	 * 获取Excel读取器，通过调用{@link ExcelReader}的read或readXXX方法读取Excel内容<br>
 	 * 默认调用第一个sheet，读取结束自动关闭流
@@ -231,7 +227,7 @@ public class ExcelUtil {
 	public static ExcelReader getReader(InputStream bookStream, boolean closeAfterRead) {
 		return getReader(bookStream, 0, closeAfterRead);
 	}
-	
+
 	/**
 	 * 获取Excel读取器，通过调用{@link ExcelReader}的read或readXXX方法读取Excel内容<br>
 	 * 读取结束自动关闭流
@@ -256,7 +252,7 @@ public class ExcelUtil {
 	public static ExcelReader getReader(InputStream bookStream, int sheetIndex, boolean closeAfterRead) {
 		return new ExcelReader(bookStream, sheetIndex, closeAfterRead);
 	}
-	
+
 	/**
 	 * 获取Excel读取器，通过调用{@link ExcelReader}的read或readXXX方法读取Excel内容<br>
 	 * 读取结束自动关闭流
@@ -282,114 +278,6 @@ public class ExcelUtil {
 	}
 
 	/**
-	 * 加载工作簿
-	 * 
-	 * @param excelFilePath Excel文件路径，绝对路径或相对于ClassPath路径
-	 * @return {@link Workbook}
-	 * @since 3.1.1
-	 */
-	public static Workbook loadBook(String excelFilePath) {
-		return loadBook(FileUtil.file(excelFilePath), null);
-	}
-
-	/**
-	 * 加载工作簿
-	 * 
-	 * @param excelFile Excel文件
-	 * @return {@link Workbook}
-	 */
-	public static Workbook loadBook(File excelFile) {
-		return loadBook(excelFile, null);
-	}
-	
-	/**
-	 * 加载工作簿，只读模式
-	 * 
-	 * @param excelFile Excel文件
-	 * @param password Excel工作簿密码，如果无密码传{@code null}
-	 * @return {@link Workbook}
-	 */
-	public static Workbook loadBook(File excelFile, String password) {
-		return loadBook(FileUtil.getInputStream(excelFile), password, true);
-	}
-	
-	/**
-	 * 加载工作簿
-	 * 
-	 * @param in Excel输入流
-	 * @param closeAfterRead 读取结束是否关闭流
-	 * @return {@link Workbook}
-	 */
-	public static Workbook loadBook(InputStream in, boolean closeAfterRead) {
-		return loadBook(in, null, closeAfterRead);
-	}
-
-	/**
-	 * 加载工作簿
-	 * 
-	 * @param in Excel输入流
-	 * @param password 密码
-	 * @param closeAfterRead 读取结束是否关闭流
-	 * @return {@link Workbook}
-	 * @since 4.0.3
-	 */
-	public static Workbook loadBook(InputStream in, String password, boolean closeAfterRead) {
-		try {
-			return WorkbookFactory.create(in, password);
-		} catch (Exception e) {
-			throw new POIException(e);
-		} finally {
-			if(closeAfterRead) {
-				IoUtil.close(in);
-			}
-		}
-	}
-
-	/**
-	 * 根据文件路径创建新的工作簿，文件路径
-	 * 
-	 * @param destFilePath 目标文件路径，文件可以不存在，通过扩展名判断Workbook类型
-	 * @return {@link Workbook}
-	 * @since 3.2.0
-	 */
-	public static Workbook createBook(String destFilePath) {
-		Workbook workbook;
-		if (StrUtil.endWithIgnoreCase(destFilePath, "xlsx")) {
-			workbook = new XSSFWorkbook();
-		} else {
-			workbook = new HSSFWorkbook();
-		}
-		return workbook;
-	}
-
-	/**
-	 * 根据文件路径创建新的工作簿
-	 * 
-	 * @param destFile 目标文件，文件可以不存在
-	 * @return {@link Workbook}
-	 * @since 3.2.0
-	 */
-	public static Workbook createBook(File destFile) {
-		return createBook(destFile.getName());
-	}
-
-	/**
-	 * 将Excel Workbook刷出到输出流
-	 * 
-	 * @param book {@link Workbook}
-	 * @param out 输出流
-	 * @throws IORuntimeException IO异常
-	 * @since 3.2.0
-	 */
-	public static void writeBook(Workbook book, OutputStream out) throws IORuntimeException {
-		try {
-			book.write(out);
-		} catch (IOException e) {
-			throw new IORuntimeException(e);
-		}
-	}
-	
-	/**
 	 * 获取或者创建sheet表<br>
 	 * 如果sheet表在Workbook中已经存在，则获取之，否则创建之
 	 * 
@@ -399,12 +287,12 @@ public class ExcelUtil {
 	 * @since 4.0.2
 	 */
 	public static Sheet getOrCreateSheet(Workbook book, String sheetName) {
-		if(null == book) {
+		if (null == book) {
 			return null;
 		}
 		sheetName = StrUtil.isBlank(sheetName) ? "sheet1" : sheetName;
 		Sheet sheet = book.getSheet(sheetName);
-		if(null == sheet) {
+		if (null == sheet) {
 			sheet = book.createSheet(sheetName);
 		}
 		return sheet;